refactor: move plot_styles to tools
