+++ simulations/emp_vs_ed/nadph/plot_nadph_titration.py
@@ -21,9 +21,9 @@ REPO_ROOT = FILE_PATH.parents[2]
 OUT_PATH = FILE_PATH / "results"
 OUT_PATH.mkdir(exist_ok=True)
 
-# Import plot_styles.py from the root of the repo
+# Import the shared plot styles from tools/
 sys.path.append(str(REPO_ROOT))
-from plot_styles import set_plot_style, summer_colors  # noqa: E402
+from tools.plot_styles import set_plot_style, summer_colors  # noqa: E402
 
 # NADPH source columns in the order/colors we want to show them
 SOURCE_COLS = [
