+++ simulations/emp_vs_ed/tradeoff_on_different_substrates/plot_forced_routing.py
@@ -10,9 +10,9 @@ REPO_ROOT = FILE_PATH.parents[2]
 OUT_PATH = FILE_PATH / "results"
 OUT_PATH.mkdir(exist_ok=True)
 
-# Import plot_styles.py from the root of the repo
+# Import the shared plot styles from tools/
 sys.path.append(str(REPO_ROOT))
-from plot_styles import set_plot_style, summer_colors
+from tools.plot_styles import set_plot_style, summer_colors
 
 # Set some parameters for the plot
 width = 0.35
